pyproject.toml
[project]
name = "homelab-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aio-pika>=9.5.5",
    "aiohttp>=3.11.16",
    "aiormq>=6.8.1",
    "boto3>=1.37.30",
    "cryptography>=44.0.2",
    "influxdb-client>=1.48.0",
    "minio>=7.2.15",
    "pika>=1.3.2",
    "prometheus-client>=0.21.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "schedule>=1.2.2",
    "zstandard>=0.23.0",
]
